Improve error handling for ParserYaml
Instead of printing to stderr or throwing random exceptions, ParserYaml
now should return all errors it encounters in a vector. If no errors are
found, an empty optional is returned instead. Parsing continues even if
an error is encountered in order to catch all errors in a single call,
the provided object will still be modified, but we cannot assure what
its state will be on return, so the caller should check for errors and
discard the message if any are found.